Update README.md
Linking Metamask with Ethereum platforms ,Explanation: It would be much better if Metamask, which is the first wallet in the chain border, is automatically linked to the app and Ethereum verification platforms that the user needs.  It pays for itself, and it pays off sooner for all chains and cryptocurrencies.